requirements.txt
agno>=0.1.0
python-dotenv>=0.19.0
typer>=0.9.0
rich>=10.0.0
typing-extensions>=4.0.0
openai>=1.0.0
python-telegram-bot>=20.0 